requirements.txt
tortoise-orm[asyncpg]
fastapi
pydantic
uvicorn

git+https://github.com/HexChap/MSCore#egg=ms_core